+++ py_pdf_parser/exceptions.py
@@ -20,6 +20,10 @@ class MultipleElementsFoundError(PDFParserError):
     pass
 
 
+class ElementOutOfRangeError(PDFParserError):
+    pass
+
+
 # Sectioning
 class InvalidSectionError(PDFParserError):
     pass